@@ -2,41 +2,56 @@
     "type": "project",
     "license": "proprietary",
     "require": {
-        "php": ">=7.1.3",
+        "php": ">=8.1",
         "ext-ctype": "*",
         "ext-iconv": "*",
         "ext-json": "*",
-        "doctrine/doctrine-migrations-bundle": "1.*",
-        "sensio/framework-extra-bundle": "5.*",
-        "symfony/asset": "4.*",
-        "symfony/console": "4.*",
-        "symfony/event-dispatcher": "4.*",
-        "symfony/flex": "1.*",
-        "symfony/framework-bundle": "4.*",
-        "symfony/orm-pack": "1.*",
-        "symfony/serializer-pack": "1.*",
-        "symfony/translation": "4.*",
-        "symfony/twig-bundle": "4.*",
+        "babdev/pagerfanta-bundle": "3.*",
+        "doctrine/annotations": "^1.0",
+        "doctrine/doctrine-bundle": "^2.6",
+        "doctrine/doctrine-migrations-bundle": "3.*",
+        "doctrine/orm": "^2.12",
+        "phpdocumentor/reflection-docblock": "^5.3",
+        "phpstan/phpdoc-parser": "^1.5",
+        "sensio/framework-extra-bundle": "6.*",
+        "symfony/asset": "6.1.*",
+        "symfony/console": "6.1.*",
+        "symfony/dotenv": "6.1.*",
+        "symfony/event-dispatcher": "6.*",
+        "symfony/flex": "2.*",
+        "symfony/framework-bundle": "6.1.*",
+        "symfony/property-access": "6.1.*",
+        "symfony/property-info": "6.1.*",
+        "symfony/proxy-manager-bridge": "6.1.*",
+        "symfony/serializer": "6.1.*",
+        "symfony/translation": "6.*",
+        "symfony/twig-bundle": "6.*",
+        "symfony/uid": "6.1.*",
         "symfony/webpack-encore-bundle": "1.*",
-        "symfony/yaml": "4.*",
-        "white-october/pagerfanta-bundle": "1.*"
+        "symfony/yaml": "6.*"
     },
     "require-dev": {
         "doctrine/data-fixtures": "1.*",
         "doctrine/doctrine-fixtures-bundle": "3.*",
-        "nelmio/alice": "3.*",
         "phpmetrics/phpmetrics": "2.*",
+        "phpunit/phpunit": "^9.5",
         "roave/security-advisories": "dev-master",
-        "squizlabs/php_codesniffer": "3.*",
-        "symfony/dotenv": "4.*",
-        "symfony/test-pack": "1.*",
-        "symfony/var-dumper": "4.*"
+        "symfony/browser-kit": "6.1.*",
+        "symfony/css-selector": "6.1.*",
+        "symfony/phpunit-bridge": "^6.1",
+        "symfony/var-dumper": "6.*"
     },
     "config": {
+        "discard-changes": true,
+        "optimize-autoloader": true,
         "preferred-install": {
             "*": "dist"
         },
-        "sort-packages": true
+        "sort-packages": true,
+        "allow-plugins": {
+            "composer/package-versions-deprecated": true,
+            "symfony/flex": true
+        }
     },
     "autoload": {
         "psr-4": {
@@ -49,8 +64,10 @@
         }
     },
     "replace": {
+        "paragonie/random_compat": "2.*",
         "symfony/polyfill-ctype": "*",
         "symfony/polyfill-iconv": "*",
+        "symfony/polyfill-php72": "*",
         "symfony/polyfill-php71": "*",
         "symfony/polyfill-php70": "*",
         "symfony/polyfill-php56": "*"
@@ -76,7 +93,8 @@
     },
     "extra": {
         "symfony": {
-            "allow-contrib": false
+            "allow-contrib": false,
+            "require": "6.1.*"
         }
     }
 }
