@@ -1,30 +1,26 @@
 <?php
+
 declare(strict_types=1);
 
 namespace App\Controller;
 
 use App\Provider\RssFeedProvider;
-use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\HttpFoundation\Response;
 
-/**
- * Class RssController
- * @package App\Controller
- */
-class RssController extends Controller
+class RssController extends AbstractController
 {
-    /**
-     * @Route("/api/feed", name="rss")
-     * @param Request $request
-     * @param RssFeedProvider $provider
-     * @return Response
-     */
-    public function __invoke(Request $request, RssFeedProvider $provider)
+    #[Route(path: '/api/feed', name: 'rss')]
+    public function __invoke(Request $request, RssFeedProvider $provider): Response
     {
-        $response = new Response($provider->getFeed());
-        $response->headers->set('Content-Type', 'application/rss+xml');
-        return $response;
+        return new Response(
+            $provider->getFeed(),
+            Response::HTTP_OK,
+            [
+                'Content-Type' => 'application/rss+xml',
+            ]
+        );
     }
 }
