@@ -3,11 +3,7 @@ declare(strict_types=1);
 
 namespace App\Events;
 
-/**
- * Class RatingEvents
- * @package App\Events
- */
 class RatingEvents
 {
-    const RATING_CREATE = 'app.event.rating.create';
+    public const RATING_CREATE = 'app.event.rating.create';
 }
